third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/declared/getAll.tentative-expected.txt

This is a testharness.js-based test.
PASS Calling StylePropertyMap.getAll with an unsupported property throws a TypeError
PASS Calling StylePropertyMap.getAll with a property not in the property model returns an empty list
PASS Calling StylePropertyMap.getAll with a custom property not in the property model returns an empty list
PASS Calling StylePropertyMap.getAll with a valid property returns a single element list with the correct entry
PASS StylePropertyMap.getAll is case-insensitive
FAIL Calling StylePropertyMap.getAll with a valid custom property returns a single element list with the correct entry assert_equals: expected "auto" but got " auto"
PASS Calling StylePropertyMap.getAll with a list-valued property returns all the values
Harness: the test ran to completion.
